Edited html file
modified:   sample_programs/joes_website/src/MainPage.html

Added some text to help with styling and formating the page. No Styling
or formatting was added yet, except for the formatting of text with h
tags and p tags.

Program was run and reviewed, and everything is running as expected at this point.